// util/interfaces.ts
export interface IDBSummoner {
    accountId: string;
    lastUpdated?: string;
    name: string;
    profileIconId: number;
    puuid: string;
    revisionDate: string;
    summonerId: string;
    summonerLevel: number;
}